// src/interfaces/IForm.tsx
export interface IForms {
  nome: string
  email: string
  telefone: string
  endereco: string
  interiores: boolean
  comercial: boolean
  arquitetonico_residencial: boolean
  legalizacao: boolean
  comprando: boolean
  recebendo_chaves: boolean
  sem_pressa: boolean
  tenho_pressa: boolean
  terreno_vazio: boolean
  imovel_antigo: boolean
  recem_entregue: boolean
  habitando: boolean
  instagram: boolean
  indicacao: boolean
  visitei: boolean
  outros: boolean
  metros_imoveis: string
  seu_objetivo: string
  quais_ambientes: string
  images: File[]
}